LineBotGroupLog/requirements.txt
Flask==2.2.2
requests==2.28.2
line-bot-sdk==2.4.2
gunicorn==20.1.0
